+++ src/Filter/FilterIn.php
@@ -3,11 +3,12 @@
 namespace Lmc\ApiFilter\Filter;
 
 use Assert\Assertion;
+use Lmc\ApiFilter\Constant\Filter;
 use Lmc\ApiFilter\Entity\Value;
 
 class FilterIn extends AbstractFilter
 {
-    public const TITLE = 'in';
+    public const TITLE = Filter::IN;
 
     public function __construct(string $column, Value $value, string $title = self::TITLE)
     {