+++ tests/Service/QueryParametersParserTest.php
@@ -15,7 +15,9 @@ class QueryParametersParserTest extends AbstractTestCase
 
     protected function setUp(): void
     {
-        $this->queryParametersParser = new QueryParametersParser();
+        $this->queryParametersParser = new QueryParametersParser(
+            new FilterFactory()
+        );
     }
 
     /**
@@ -36,29 +38,36 @@ class QueryParametersParserTest extends AbstractTestCase
         return [
             // queryParameters, expectedFilters
             'empty' => [[], []],
-            'simple eq' => [
+            'simple - implicit eq' => [
                 ['title' => 'foo'],
                 [new FilterWithOperator('title', new Value('foo'), '=', 'eq')],
             ],
-            'two cols eq' => [
+            'two cols - implicit eq' => [
                 ['title' => 'foo', 'value' => 'bar'],
                 [
                     new FilterWithOperator('title', new Value('foo'), '=', 'eq'),
                     new FilterWithOperator('value', new Value('bar'), '=', 'eq'),
                 ],
             ],
-            'eq by array' => [
+            'implicit EQ + explicit filter' => [
+                ['name' => 'Jon', 'age' => ['gt' => 20]],
+                [
+                    new FilterWithOperator('name', new Value('Jon'), '=', 'eq'),
+                    new FilterWithOperator('age', new Value(20), '>', 'gt'),
+                ],
+            ],
+            'explicit eq' => [
                 ['title' => ['eq' => 'foo']],
                 [new FilterWithOperator('title', new Value('foo'), '=', 'eq')],
             ],
-            'one col more filters' => [
+            'one col more filters - explicit' => [
                 ['title' => ['eq' => 'foo', 'gt' => 'abc']],
                 [
                     new FilterWithOperator('title', new Value('foo'), '=', 'eq'),
                     new FilterWithOperator('title', new Value('abc'), '>', 'gt'),
                 ],
             ],
-            'one col more filters + other col' => [
+            'one col more filters + other col - explicit/implicit' => [
                 ['title' => ['gt' => '0', 'lt' => '10'], 'value' => 'foo'],
                 [
                     new FilterWithOperator('title', new Value('0'), '>', 'gt'),
@@ -66,27 +75,27 @@ class QueryParametersParserTest extends AbstractTestCase
                     new FilterWithOperator('value', new Value('foo'), '=', 'eq'),
                 ],
             ],
-            'one col min max' => [
+            'one col - between - explicit' => [
                 ['title' => ['gte' => '0', 'lte' => '10']],
                 [
                     new FilterWithOperator('title', new Value('0'), '>=', 'gte'),
-                    new FilterWithOperator('title', new Value('10'), '<=', 'lt'),
+                    new FilterWithOperator('title', new Value('10'), '<=', 'lte'),
                 ],
             ],
-            'in array' => [
+            'explicit in' => [
                 ['color' => ['in' => ['red', 'green', 'blue']]],
                 [
                     new FilterIn('color', new Value(['red', 'green', 'blue'])),
                 ],
             ],
-            'eq + in array' => [
+            'implicit eq + explicit in' => [
                 ['allowed' => true, 'id' => ['in' => [1, 2, 3]]],
                 [
                     new FilterWithOperator('allowed', new Value(true), '=', 'eq'),
                     new FilterIn('id', new Value([1, 2, 3])),
                 ],
             ],
-            'tuple - eq + in array' => [
+            'tuple - implicit eq + explicit in' => [
                 ['(zone,bucket)' => '(lmc,all)', 'id' => ['in' => [1, 2, 3]]],
                 [
                     new FilterWithOperator('zone', new Value('lmc'), '=', 'eq'),
@@ -94,13 +103,61 @@ class QueryParametersParserTest extends AbstractTestCase
                     new FilterIn('id', new Value([1, 2, 3])),
                 ],
             ],
-            'tuple - between' => [
+            'tuple - between - explicit in values' => [
                 ['(number,alpha)' => ['gte' => '(0, a)', 'lt' => '(10, z)']],
                 [
                     new FilterWithOperator('number', new Value('0'), '>=', 'gte'),
                     new FilterWithOperator('alpha', new Value('a'), '>=', 'gte'),
                     new FilterWithOperator('number', new Value('10'), '<', 'lt'),
                     new FilterWithOperator('alpha', new Value('z'), '<', 'lt'),
+                ],
+            ],
+            'tuple - between - explicit in columns' => [
+                ['(age[gt],age[lt])' => '(18, 30)'],
+                [
+                    new FilterWithOperator('age', new Value(18), '>', 'gt'),
+                    new FilterWithOperator('age', new Value(30), '<', 'lt'),
+                ],
+            ],
+            'tuple - implicit eq + in' => [
+                ['(name,size)' => '(foo, [DD; D])'],
+                [
+                    new FilterWithOperator('name', new Value('foo'), '=', 'eq'),
+                    new FilterIn('size', new Value(['DD', 'D'])),
+                ],
+            ],
+            'ints - between - explicit' => [
+                ['age' => ['gt' => 18, 'lt' => 30]],
+                [
+                    new FilterWithOperator('age', new Value(18), '>', 'gt'),
+                    new FilterWithOperator('age', new Value(30), '<', 'lt'),
+                ],
+            ],
+            'explicit between + explicit in' => [
+                ['age' => ['gt' => 18, 'lt' => 30], 'size' => ['in' => ['DD', 'D']]],
+                [
+                    new FilterWithOperator('age', new Value(18), '>', 'gt'),
+                    new FilterWithOperator('age', new Value(30), '<', 'lt'),
+                    new FilterIn('size', new Value(['DD', 'D'])),
+                ],
+            ],
+            'tuple - explicit between + implicit in and eq in columns' => [
+                ['(age[gte], age[lt], size, name)' => '(18, 30, [DD; D], "Jon Snow")'],
+                [
+                    new FilterWithOperator('age', new Value(18), '>=', 'gte'),
+                    new FilterWithOperator('age', new Value(30), '<', 'lt'),
+                    new FilterIn('size', new Value(['DD', 'D'])),
+                    new FilterWithOperator('name', new Value('Jon Snow'), '=', 'eq'),
+                ],
+            ],
+            'tuple - explicit between + implicit in and eq in columns + other implicit eq' => [
+                ['(age[gte], age[lt], size, name)' => '(18, 30, [DD; D], "Jon Snow")', 'version' => 'latest'],
+                [
+                    new FilterWithOperator('age', new Value(18), '>=', 'gte'),
+                    new FilterWithOperator('age', new Value(30), '<', 'lt'),
+                    new FilterIn('size', new Value(['DD', 'D'])),
+                    new FilterWithOperator('name', new Value('Jon Snow'), '=', 'eq'),
+                    new FilterWithOperator('version', new Value('latest'), '=', 'eq'),
                 ],
             ],
         ];
@@ -123,7 +180,7 @@ class QueryParametersParserTest extends AbstractTestCase
     public function provideInvalidQueryParameters(): array
     {
         return [
-            // queryParameters
+            // queryParameters, expected message
             'empty filter' => [
                 ['column' => ['' => 'value']],
                 'Filter "" is not implemented. For column "column" with value "value".',
@@ -132,13 +189,29 @@ class QueryParametersParserTest extends AbstractTestCase
                 ['column' => ['unknown' => 'value']],
                 'Filter "unknown" is not implemented. For column "column" with value "value".',
             ],
-            'invalid tuple - too much values' => [
-                ['(id,name)' => ['eq' => '(42,foo,bar)']],
-                'Invalid tuple given - expected 2 items but parsed 3 items from "(42,foo,bar)".',
+            'function' => [
+                ['function' => '(arg1, arg2)'],
+                'Invalid combination of a tuple and a scalar. Column function and value (arg1, arg2).',
             ],
-            'invalid tuple - insufficient values' => [
-                ['(id, name, type)' => ['eq' => '(42, foo)']],
-                'Invalid tuple given - expected 3 items but parsed 2 items from "(42, foo)".',
+            'tuple columns and a single value' => [
+                ['(col1, col2)' => 'value'],
+                'Invalid combination of a tuple and a scalar. Column (col1, col2) and value value.',
+            ],
+            'more columns than values' => [
+                ['(col1, col2, col3)' => '(val1, val2)'],
+                'Number of given columns (3) and values (2) in tuple are not same.',
+            ],
+            'more values than columns' => [
+                ['(col1, col2)' => '(val1, val2, val3)'],
+                'Number of given columns (2) and values (3) in tuple are not same.',
+            ],
+            'invalid tuple - explicit filters' => [
+                ['(id,name)' => ['eq' => '(42,foo,bar)']],
+                'Number of given columns (2) and values (3) in tuple are not same.',
+            ],
+            'invalid tuple - filter definition in columns and values' => [
+                ['(first[gt],second[lt])' => ['eq' => '(1,2)']],
+                'Filters can be specified either in columns or in values - not in both',
             ],
             'tuples in IN filter' => [
                 ['(id, name)' => ['in' => ['(1,one)', '(2,two)']]],